- move event call hooks to ForgeHooksClient - Streamline Font override handling - fix tooltip stack missing from events on creative search tab
