+++ src/main/java/net/minecraftforge/client/MinecraftForgeClient.java
@@ -20,10 +20,8 @@
 package net.minecraftforge.client;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
@@ -31,24 +29,13 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.datafixers.util.Either;
-import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.locale.Language;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.FormattedText;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
-import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.client.event.RenderTooltipEvent;
-import net.minecraftforge.common.MinecraftForge;
 import org.apache.commons.lang3.tuple.Pair;
 
 import com.google.common.cache.CacheBuilder;
@@ -210,82 +197,6 @@ public class MinecraftForgeClient
     {
         var factory = tooltipComponentFactories.get(component.getClass());
         return factory == null ? null : factory.apply(component);
-    }
-
-    public static RenderTooltipEvent.Pre preTooltipEvent(@Nonnull ItemStack stack, PoseStack matrixStack, int x, int y, int screenWidth, int screenHeight, @Nonnull Font font, @Nonnull List<ClientTooltipComponent> components)
-    {
-        var preEvent = new RenderTooltipEvent.Pre(stack, matrixStack, x, y, screenWidth, screenHeight, font, components);
-        MinecraftForge.EVENT_BUS.post(preEvent);
-        return preEvent;
-    }
-
-    public static RenderTooltipEvent.Color colorTooltipEvent(@Nonnull ItemStack stack, PoseStack matrixStack, int x, int y, @Nonnull Font font, @Nonnull List<ClientTooltipComponent> components)
-    {
-        var colorEvent = new RenderTooltipEvent.Color(stack, matrixStack, x, y, font, 0xf0100010, 0x505000FF, 0x5028007f, components);
-        MinecraftForge.EVENT_BUS.post(colorEvent);
-        return colorEvent;
-    }
-
-    public static List<ClientTooltipComponent> gatherTooltipComponents(ItemStack stack, List<? extends FormattedText> textElements, int mouseX, int screenWidth, int screenHeight, Font font)
-    {
-        return gatherTooltipComponents(stack, textElements, Optional.empty(), mouseX, screenWidth, screenHeight, font);
-    }
-
-    public static List<ClientTooltipComponent> gatherTooltipComponents(ItemStack stack, List<? extends FormattedText> textElements, Optional<TooltipComponent> itemComponent, int mouseX, int screenWidth, int screenHeight, Font font)
-    {
-        List<Either<FormattedText, TooltipComponent>> elements = textElements.stream()
-                .map((Function<FormattedText, Either<FormattedText, TooltipComponent>>) Either::left)
-                .collect(Collectors.toCollection(ArrayList::new));
-        itemComponent.ifPresent(c -> elements.add(Either.right(c)));
-
-        var event = new RenderTooltipEvent.GatherComponents(stack, screenWidth, screenHeight, elements, -1);
-        MinecraftForge.EVENT_BUS.post(event);
-        if (event.isCanceled()) return List.of();
-
-        // text wrapping
-        int tooltipTextWidth = event.getTooltipElements().stream()
-                .mapToInt(either -> either.map(font::width, component -> 0))
-                .max()
-                .orElse(0);
-
-        boolean needsWrap = false;
-
-        int tooltipX = mouseX + 12;
-        if (tooltipX + tooltipTextWidth + 4 > screenWidth)
-        {
-            tooltipX = mouseX - 16 - tooltipTextWidth;
-            if (tooltipX < 4) // if the tooltip doesn't fit on the screen
-            {
-                if (mouseX > screenWidth / 2)
-                    tooltipTextWidth = mouseX - 12 - 8;
-                else
-                    tooltipTextWidth = screenWidth - 16 - mouseX;
-                needsWrap = true;
-            }
-        }
-
-        if (event.getMaxWidth() > 0 && tooltipTextWidth > event.getMaxWidth())
-        {
-            tooltipTextWidth = event.getMaxWidth();
-            needsWrap = true;
-        }
-
-        int tooltipTextWidthF = tooltipTextWidth;
-        if (needsWrap)
-        {
-            return event.getTooltipElements().stream()
-                    .flatMap(either -> either.map(
-                            text -> font.split(text, tooltipTextWidthF).stream().map(ClientTooltipComponent::create),
-                            component -> Stream.of(ClientTooltipComponent.create(component))
-                    ))
-                    .toList();
-        }
-        return event.getTooltipElements().stream()
-                .map(either -> either.map(
-                        text -> ClientTooltipComponent.create(text instanceof Component ? ((Component) text).getVisualOrderText() : Language.getInstance().getVisualOrder(text)),
-                        ClientTooltipComponent::create
-                ))
-                .toList();
     }
 
 }
